+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/planner/login/Planner.java
@@ -1,4 +1,4 @@
-package com.mysite.weddingyou_backend.planner;
+package com.mysite.weddingyou_backend.planner.login;
 
 import java.time.LocalDateTime;
