+++ weddingyou_backend/src/main/java/com/mysite/weddingyou_backend/user/login/UserRepository.java
@@ -1,4 +1,4 @@
-package com.mysite.weddingyou_backend.planner;
+package com.mysite.weddingyou_backend.user.login;
 
 import java.util.Optional;
 
@@ -7,19 +7,19 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.mysite.weddingyou_backend.planner.Planner;
+import com.mysite.weddingyou_backend.user.login.User;
 
 import jakarta.transaction.Transactional;
 
 
 @Repository
-public interface PlannerRepository extends JpaRepository<Planner, Long> {
-	//이메일로 회원 정보 조회(select * from Planner where email=?)
-	Planner findByEmail(String Email); //optional은 null방지..?
+public interface UserRepository extends JpaRepository<User, Long> {
+	//이메일로 회원 정보 조회(select * from User where email=?)
+	User findByEmail(String Email); //optional은 null방지..?
 	
 	//insert, update, delete 할때 필요함
 	@Modifying
 	@Transactional
-	@Query(value="update planner set password = :password where email = :email", nativeQuery=true)
+	@Query(value="update user set password = :password where email = :email", nativeQuery=true)
 	public int updatePassword(String email, String password);
 }
